+++ tests/test_board.py
@@ -12,62 +12,6 @@ class TestBoard(unittest.TestCase):
         self.assertEqual(len(board.grid),15)
         self.assertEqual(len(board.grid[0]),15)
 
-    def test_simple(self):
-        cell_1 = Cell(multiplier=1,multiplier_type="letter")
-        cell_1.add_letter(Tile('C', 1)) 
-        cell_2 = Cell(multiplier=1,multiplier_type="letter")
-        cell_2.add_letter(Tile('A', 1))
-        cell_3 = Cell(multiplier=1,multiplier_type="letter")
-        cell_3.add_letter(Tile('S', 2))
-        cell_4 = Cell(multiplier=1,multiplier_type="letter")
-        cell_4.add_letter(Tile('A', 1))
-
-        word = [cell_1, cell_2, cell_3, cell_4]
-        value = Board().calculate_word_value(word)
-        self.assertEqual(value, 5)
-    
-    def test_letter_multiplier(self):
-        cell_1 = Cell(multiplier=1,multiplier_type="letter")
-        cell_1.add_letter(Tile('C', 1)) 
-        cell_2 = Cell(multiplier=1,multiplier_type="letter")
-        cell_2.add_letter(Tile('A', 1))
-        cell_3 = Cell(multiplier=2,multiplier_type="letter")
-        cell_3.add_letter(Tile('S', 2))
-        cell_4 = Cell(multiplier=1,multiplier_type="letter")
-        cell_4.add_letter(Tile('A', 1))
-
-        word = [cell_1, cell_2, cell_3, cell_4]
-        value = Board().calculate_word_value(word)
-        self.assertEqual(value, 7)
-    
-    def test_word_multiplier(self):
-        cell_1 = Cell(multiplier=1,multiplier_type="letter")
-        cell_1.add_letter(Tile('C', 1)) 
-        cell_2 = Cell(multiplier=1,multiplier_type="letter")
-        cell_2.add_letter(Tile('A', 1))
-        cell_3 = Cell(multiplier=2,multiplier_type="word")
-        cell_3.add_letter(Tile('S', 2))
-        cell_4 = Cell(multiplier=1,multiplier_type="letter")
-        cell_4.add_letter(Tile('A', 1))
-
-        word = [cell_1, cell_2, cell_3, cell_4]
-        value = Board().calculate_word_value(word)
-        self.assertEqual(value, 10)
-
-    def test_letter_and_word_multiplier(self):
-        cell_1 = Cell(multiplier=2,multiplier_type="letter")
-        cell_1.add_letter(Tile('C', 1)) 
-        cell_2 = Cell(multiplier=1,multiplier_type="letter")
-        cell_2.add_letter(Tile('A', 1))
-        cell_3 = Cell(multiplier=2,multiplier_type="word")
-        cell_3.add_letter(Tile('S', 2))
-        cell_4 = Cell(multiplier=1,multiplier_type="letter")
-        cell_4.add_letter(Tile('A', 1))
-
-        word = [cell_1, cell_2, cell_3, cell_4]
-        value = Board().calculate_word_value(word)
-        self.assertEqual(value, 12)
-
     @patch('game.board.dle.search_by_word')
     def test_len_of_word_in_board_x(self,mock_search_by_word):
         mock_search_by_word.return_value.title = 'facultad | Definición | Diccionario de la lengua española | RAE - ASALE'
@@ -310,8 +254,203 @@ O   3W|  |  |2L|  |  |  |3W|  |  |  |2L|  |  |3W|
         self.assertEqual(board.grid[9][7].letter.letter, "S")
         self.assertEqual(board.grid[10][7].letter.letter, "A")
 
+    def test_get_word_without_intersections_intersected(self):
+        board = Board()
+        board.grid[7][6].letter = Tile('C',3)
+        board.grid[7][7].letter = Tile('A',1)
+        board.grid[7][8].letter = Tile('S',1)
+        board.grid[7][9].letter = Tile('A',1)
+        word = 'laso'
+        location = (6,7)
+        orientation = "V"
+        result_word = board.get_word_without_intersections(word,location,orientation)
+        self.assertEqual(result_word,'lso')
 
+    def test_get_word_without_intersections_intersected(self):
+        board = Board()
+        board.grid[7][7].letter = Tile('L',1)
+        board.grid[7][8].letter = Tile('A',1)
+        word = 'las'
+        location = (7,7)
+        orientation = "H"
+        result_word = board.get_word_without_intersections(word,location,orientation)
+        self.assertEqual(result_word,'s')
 
+class TestCalculateWordValue(unittest.TestCase):
+    def test_simple(self):
+        board = Board()        
+        word = 'casa'
+        pos=(10,6)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 6)
 
+    def test_simple_with_double_letter(self):
+        board = Board()
+        word = 'lluvia'
+        pos=(10,5)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 15)
+        
+    def test_with_word_multiplier_2(self):
+        board = Board()        
+        word = 'casa'
+        pos=(7,6)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 7)
+
+    def test_with_word_multiplier_3(self):
+        board = Board()        
+        word = 'casa'
+        pos=(14,6)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 8)
+
+    def test_with_letter_multiplier_2(self):
+        board = Board()        
+        word = 'casa'
+        pos=(8,8)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 9)
+
+    def test_with_letter_multiplier_3(self):
+        board = Board()        
+        word = 'casa'
+        pos=(9,8)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 8)
+
+    def test_with_letter_multiplier_and_word_multiplier(self):
+        board = Board()        
+        word = 'casa'
+        pos=(14,0)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 13)
+
+    def test_with_letter_multiplier_and_word_multiplier_no_active(self):
+        board = Board()        
+        word = 'casa'
+        pos=(14,0)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        value = board.calculate_word_value(word, pos, horizontal=True)
+        self.assertEqual(value, 6)
+
+    def test_simple_intersection(self):
+        board = Board()
+        board.grid[7][2].tile = Tile('E', 1)
+        board.grid[7][3].tile = Tile('S', 1)
+        board.grid[7][4].tile = Tile('T', 1)
+        board.grid[7][5].tile = Tile('R', 1)
+        board.grid[7][6].tile = Tile('E', 1)
+        board.grid[7][7].tile = Tile('N', 1)
+        board.grid[7][8].tile = Tile('O', 1)
+        board.grid[7][7].active = False
+        word = 'mes'
+        pos=(6,6)
+        horizontal = True
+        value = board.calculate_word_value(word, pos, horizontal)
+        self.assertEqual(value, 9)
+
+    def test_simple_intersection_inverted(self):
+        board = Board()
+        board.grid[7][2].tile = Tile('E', 1)
+        board.grid[7][3].tile = Tile('S', 1)
+        board.grid[7][4].tile = Tile('T', 1)
+        board.grid[7][5].tile = Tile('R', 1)
+        board.grid[7][6].tile = Tile('E', 1)
+        board.grid[7][7].tile = Tile('N', 1)
+        board.grid[7][8].tile = Tile('O', 1)
+        board.grid[7][7].active = False
+        word = 'mes'
+        pos=(8,6)
+        horizontal = True
+        value = board.calculate_word_value(word, pos, horizontal)
+        self.assertEqual(value, 9)
+
+    def test_simple_intersection_2(self):
+        board = Board()
+        board.grid[7][5].letter = Tile('C', 3)
+        board.grid[7][6].letter = Tile('A', 1)
+        board.grid[7][7].letter = Tile('S', 1)
+        board.grid[7][8].letter = Tile('A', 1)
+        board.grid[6][8].letter = Tile('L', 1)
+        board.grid[8][8].letter = Tile('S', 1)
+        board.grid[9][8].letter = Tile('O', 1)
+        board.grid[9][7].letter = Tile('S', 1)
+        board.grid[9][6].letter = Tile('O', 1)
+        word = 'cosa'
+        pos=(7,5)
+        horizontal = False
+        value = board.calculate_word_value(word, pos, horizontal)
+        self.assertEqual(value, 14)
+
+    def test_simple_intersection_inverted_2(self):
+        board = Board()
+        board.grid[7][5].tile = Tile('C', 3)
+        board.grid[7][6].tile = Tile('A', 1)
+        board.grid[7][7].tile = Tile('S', 1)
+        board.grid[7][8].tile = Tile('A', 1)
+        board.grid[8][5].tile = Tile('O', 1)
+        board.grid[9][5].tile = Tile('S', 1)
+        board.grid[10][5].tile = Tile('A', 1)
+        board.grid[9][7].tile = Tile('S', 1)
+        board.grid[9][6].tile = Tile('O', 1)
+        board.grid[9][8].tile = Tile('O', 1)
+        board.grid[7][7].active = False
+        word = 'laso'
+        pos=(5,9)
+        horizontal = False
+        value = board.calculate_word_value(word, pos, horizontal)
+        self.assertEqual(value, 6)
+
+    def test_simple_intersection_inverted_3(self):
+        board = Board()
+        board.grid[7][5].tile = Tile('C', 3)
+        board.grid[7][6].tile = Tile('A', 1)
+        board.grid[7][7].tile = Tile('S', 1)
+        board.grid[7][8].tile = Tile('A', 1)
+        board.grid[8][5].tile = Tile('O', 1)
+        board.grid[9][5].tile = Tile('S', 1)
+        board.grid[10][5].tile = Tile('A', 1)
+        board.grid[9][7].tile = Tile('S', 1)
+        board.grid[9][6].tile = Tile('O', 1)
+        board.grid[9][8].tile = Tile('O', 1)
+        board.grid[7][7].active = False
+        board.grid[9][5].active = False
+        word = 'lasos'
+        pos=(5,9)
+        horizontal = False
+        value = board.calculate_word_value(word, pos, horizontal)
+        self.assertEqual(value, 9)
+
+    def test_out_of_range_side(self):
+        board = Board()
+        word = 'casa'
+        pos = (14,10)
+        horizontal = True
+        value = board.calculate_word_value(word, pos, horizontal)
+        pos = (0,0)
+        horizontal = True
+        value = board.calculate_word_value(word, pos, horizontal)
+        self.assertEqual(value!=None, True)
+
+    def test_letter_not_available(self):
+        board = Board()
+        word = 'c?sa'
+        pos = (7,7)
+        horizontal = True
+        value = board.calculate_word_value(word,pos,horizontal)
+        self.assertEqual(value, 8)
+
+    def test_calculate_word_value_with_intersection(self):
+        board = Board()
+        board.grid[7][7].letter = Tile('C',3)
+        board.grid[7][8].letter = Tile('A',1)
+        board.grid[7][9].letter = Tile('S',2)
+        board.grid[7][10].letter = Tile('A',1)
+        board.grid[7][7].active = False
+        word = 'faca'
+        location = (6,8)
+        orientation = False
+        self.assertEqual(board.calculate_word_value(word,location,orientation),16)
 if __name__ == '__main__':
     unittest.main()