+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+### [Released] - 2023-8-11
+
+### Added
+
+- Create calculate value word and his tests
+- Add validate 
+
+### Changed
+
+- Solve problems with put word in main
 
 
 ### [Development]